Resource states are readonly.
In Ember 1.x the arity of a computed property function is checked.
If it takes no arguments, it can be totally overwritten. E.g. a call
to `set` will cause the property to *never* be computed again.

However if you specify that you would like to accept two arguments
`key`, and `value`, then the behavior is similar to that of `Ember
0.9.x`.