Do not share internal data with other objects.
If we set a nested resource on an object, we do not want to take over
that object's internal data structures. We were accidentally using a
reference to the nested object's `data`. So when we we modified the
`data` of the parent object, the nested object would change as well.