Avoid deprecation warnings for shared getter and setter
Ember.js 1.12 introduced an improved syntax[1] for computed properties with
a setter.

Consequently, any CP having an arity greater than 1 is flagged as deprecated
and a warning is issued to the console:

```
DEPRECATION: Using the same function as getter and setter is deprecated. See http://emberjs.com/deprecations/v1.x/#toc_computed-properties-with-a-shared-getter-and-setter for more details.
        at ComputedProperty (http://localhost:7357/static-assets/vendor.js:20233)
        at computed (http://localhost:7357/static-assets/vendor.js:20694)
        at http://localhost:7357/static-assets/vendor.js:79924
        at http://localhost:7357/static-assets/vendor.js:80572
```

Removing the second argument on any function where the argument isn't
actually used, prevents those warnings.  Alas, it also has the unwanted
side effect of allowing to change the value of the property which simply
would have been ignored beforehand.

So, mark the property as `readOnly` which is a better solution anyway
because it prevents that an attempt for changing the value of a property
goes through undetected when it should have been forbidden.

[1]: http://emberjs.com/deprecations/v1.x/#toc_computed-properties-with-a-shared-getter-and-setter